Move JwsCompactTyped to signum
