+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/wallet/lib/jws/SdJwtSigned.kt
@@ -2,8 +2,8 @@ package at.asitplus.wallet.lib.jws
 
 import at.asitplus.KmmResult
 import at.asitplus.catching
-import at.asitplus.openid.JwsCompactTyped
 import at.asitplus.signum.indispensable.josef.JwsCompact
+import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 import at.asitplus.wallet.lib.data.KeyBindingJws
 
 /**
