+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/JwtTokenService.kt
@@ -1,7 +1,7 @@
 package at.asitplus.wallet.lib.oauth2
 
 import at.asitplus.catching
-import at.asitplus.openid.JwsCompactTyped
+import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 import at.asitplus.openid.OpenIdAuthorizationDetails
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.wallet.lib.jws.JwsContentTypeConstants
