+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/ClientIdScheme.kt
@@ -1,7 +1,7 @@
 package at.asitplus.wallet.lib.openid
 
 import at.asitplus.iso.sha256
-import at.asitplus.openid.JwsCompactTyped
+import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 import at.asitplus.openid.OpenIdConstants
 import at.asitplus.signum.indispensable.io.Base64UrlStrict
 import at.asitplus.signum.indispensable.josef.JsonWebToken
