+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciSameScopeTest.kt
@@ -14,7 +14,7 @@ package at.asitplus.wallet.lib.oidvci
 
 import at.asitplus.catching
 import at.asitplus.catchingUnwrapped
-import at.asitplus.openid.JwsCompactTyped
+import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 import at.asitplus.openid.RequestParameters
 import at.asitplus.openid.TokenResponseParameters
 import at.asitplus.testballoon.withFixtureGenerator
