+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/OpenIdRequestParserTests.kt
@@ -3,7 +3,7 @@ package at.asitplus.wallet.lib.openid
 import at.asitplus.dcapi.request.DCAPIWalletRequest
 import at.asitplus.openid.AuthenticationRequestParameters
 import at.asitplus.openid.JarRequestParameters
-import at.asitplus.openid.JwsCompactTyped
+import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 import at.asitplus.openid.RequestParametersFrom
 import at.asitplus.testballoon.invoke
 import at.asitplus.testballoon.withFixtureGenerator
