+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/VerifiablePresentationParsed.kt
@@ -1,6 +1,6 @@
 package at.asitplus.wallet.lib.data
 
-import at.asitplus.openid.JwsCompactTyped
+import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 
 
 /**
