+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/rfc/tokenStatusList/agents/StatusIssuer.kt
@@ -1,6 +1,6 @@
 package at.asitplus.wallet.lib.data.rfc.tokenStatusList.agents
 
-import at.asitplus.openid.JwsCompactTyped
+import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 import at.asitplus.signum.indispensable.cosef.CoseSigned
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.RevocationList
 import at.asitplus.wallet.lib.data.rfc.tokenStatusList.RevocationList.Kind.STATUS_LIST
